# Replace generated CIDs with LOOKUP_CID macro