# The problem

Currently, the compiler backed generates a C macro for each function name, e.g.

```c

//GENERATED_CONSTRUCTOR_IDS_START//
#define _GEN_ (31)
#define _SUM_ (28)
#define _MAIN_ (32)

//GENERATED_CONSTRUCTOR_IDS_END//
```
These macros are then used in generated case splits, e.g.
```c
         switch (fun)
          //GENERATED_REWRITE_RULES_STEP_0_START//
          {
            case _GEN_: { //<snip>
            };
            case _SUM_: { //<snip>
            };
            case _MAIN_: {//<snip>
            };

```

This causes the following issues:

* Encoding names with special characters (e.g. `IO.put_char`) requires [hard-to-get-right escaping code](https://github.com/Kindelia/HVM/blob/0bcbae286f995bc14895b0774914b1e19cf5e34e/src/compiler.rs#L32-L39) in the HVM compiler
* Downstream users of HVM may need to [re-implement this escaping code](https://github.com/lawcho/rpphp/blob/14ed49dc2e1f7a7de5e53a8c3f10107a1bfcd8da/platgen.hs#L63-L68)
* Even if the escaping is clash-free, the macro names can still clash with macros from libraries

# A solution

This PR fixes the issues above by replacing the family of `_COMPILED_NAME_` macros with a single `LOOKUP_CID` function macro, e.g.

```c
#define LOOKUP_CID(str) (\
(strcmp(str,"Gen")==0) ? 28 : \
(strcmp(str,"Main")==0) ? 32 : \
(strcmp(str,"Sum")==0) ? 31 : \
\
-1)
```
This macro is used similarly to before:
```c
          //GENERATED_REWRITE_RULES_STEP_0_START//
            if (fun == LOOKUP_CID("Gen") ) { //<snip>
              break;
            }
            if (fun == LOOKUP_CID("Main") ) { //<snip>
              break;
            }
            if (fun == LOOKUP_CID("Sum") ) { //<snip>
              break;
            }
```

# Benchmarks

In theory, this change increases the complexity of constructor-comparison from `O(1)` to `O(# constructors)`.
In practice, I did not see a noticeable slowdown on the test suite.

I think this is because the C optimizer evaluates `LOOKUP_CID(str)` fully at compile-time, when `str` is constant.

## Timings before this PR
```bash
time python3 tests/test_cli.py --run-mode compiled --hvm-cmd="hvm" >/dev/null

real    0m6,260s
user    0m4,669s
sys     0m1,645s
```
## Timings after this PR
```bash
$ time python3 tests/test_cli.py --run-mode compiled --hvm-cmd="./target/debug/hvm" >/dev/null

real    0m7,367s
user    0m5,424s
sys     0m1,613s