@lawcho 2022-10-12
Update: I've replaced the Rust-unfriendly C macro with an inline function, which gives similar performance.

I haven't seen an API for the Rust rewrite yet, but I'm guessing it will contain a function like `lookup_cid`.
If it does, then this PR brings the C runtime one step closer to the Rust version.

## Timings after this PR (commit 5d4b191)

```bash
$ time python3 tests/test_cli.py --run-mode compiled --hvm-cmd="./target/debug/hvm" >/dev/null

real	0m6,385s
user	0m4,986s
sys	0m1,344s

```