@lawcho 2022-11-22
No, this PR only touches the C backend.